unicorn: Add KernelSU

+++ vendorsetup.sh
@@ -26,6 +26,11 @@ if [ -d "kernel/xiaomi/sm8450" ]; then
     echo "kernel/xiaomi/sm8450: done"
 else
     git clone -b lineage-21 --single-branch https://github.com/dopaemon/android_kernel_xiaomi_sm8450.git kernel/xiaomi/sm8450
+fi
+
+if [ -d "kernel/xiaomi/sm8450/KernelSU" ]; then
+    echo "KernelSU: done"
+else
     cd kernel/xiaomi/sm8450
     rm -rf KernelSU
     git submodule add -f https://github.com/tiann/KernelSU.git KernelSU
